SAPA-338 rename use_stream variable as precheck
use_stream was renamed as precheck, which defaults to True. Tar archives
are now read always in stream mode and the precheck variable defines
whether or not to read the whole archive before extraction.
precheck=False saves on archive read, but leaves the cleanup for the end
user.

precheck variable was added also for zip extraction.