TPASPKT-991 Fix unit tests
The test previously asserted that a certain error text produced by
Python module zipfile was displayed. However, the error message has
changed somewhere in between Python 3.6 and 3.9, causing the test to
fail in RHEL9 environment. Use custom error message and test against
that to make the test more robust, but show the stacktrace of the error
in case the Python internal message gives some more information.